ResourceEmbedder: Fix warning

+++ ResourceEmbedder/ResourceEmbedder.cpp
@@ -146,7 +146,7 @@ void ExportFile(const char* pName, const char* pInputFilename, const char* pOutp
 
 	fprintf(pSourceFile, "#include \"%s.h\"\n", sName.c_str());
 
-	for (int iNamespace = 0; iNamespace < oNamespaces.size(); ++iNamespace)
+	for (size_t iNamespace = 0; iNamespace < oNamespaces.size(); ++iNamespace)
 	{
 		std::string sNamespace = oNamespaces[iNamespace];
 		ReplaceAll(sDefine, ".", "_");
@@ -197,7 +197,7 @@ void ExportFile(const char* pName, const char* pInputFilename, const char* pOutp
 	fprintf(pHeaderFile, "%s}\n", sIndent.c_str());
 	fprintf(pSourceFile, "%s}\n", sIndent.c_str());
 
-	for (int iNamespace = 0; iNamespace < oNamespaces.size(); ++iNamespace)
+	for (size_t iNamespace = 0; iNamespace < oNamespaces.size(); ++iNamespace)
 	{
 		std::string sNamespaceIndent(oNamespaces.size() - iNamespace - 1, '\t');
 		fprintf(pHeaderFile, "%s}\n", sNamespaceIndent.c_str());
